ajustes maria

+++ src/main/java/com/pousada/service/ReservaService.java
@@ -1,13 +1,11 @@
 package com.pousada.service;
 
-import com.pousada.domain.entity.ComodidadeEntity;
 import com.pousada.domain.entity.ReservaEntity;
-import com.pousada.domain.repository.AcomodacaoRepository;
 import com.pousada.domain.repository.ReservaRepository;
-import com.pousada.dto.ComodidadeDTO;
 import com.pousada.dto.ReservaDTO;
 import com.pousada.enums.StatusReservaEnum;
 import com.pousada.exception.AcomodacaoOcupadaException;
+import com.pousada.exception.ReservaEmAndamentoOuFinalizadaException;
 import com.pousada.exception.ReservaNaoEncontradaException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -15,8 +13,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-import java.time.temporal.ChronoUnit;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -26,7 +23,6 @@ public class ReservaService {
 
     private final ModelMapper modelMapper;
     private final ReservaRepository reservaRepository;
-    private final AcomodacaoRepository acomodacaoRepository;
 
     public ReservaDTO criarReserva(ReservaDTO reservaDTO) {
         if (existeReservaNoPeriodo(reservaDTO)) {
@@ -44,15 +40,22 @@ public class ReservaService {
         return modelMapper.map(reservaEntitySalva, ReservaDTO.class);
     }
 
-    public void deletarReservaPorId(Long id) {
+    public void deletarReservaPorId(Integer id) {
         if (reservaRepository.existsById(id)) {
+            ReservaEntity reserva = reservaRepository.findById(id)
+                    .orElseThrow(() -> new ReservaNaoEncontradaException("A reserva com o ID " + id + " não existe."));
+
+            if (Arrays.asList("EM_ANDAMENTO", "FINALIZADA").contains(String.valueOf(reserva.getStatusReserva()))) {
+                throw new ReservaEmAndamentoOuFinalizadaException("Não é possível excluir uma reserva finalizada ou em andamento.");
+            }
+
             reservaRepository.deleteById(id);
         } else {
             throw new ReservaNaoEncontradaException("A reserva com o ID " + id + " não existe.");
         }
     }
 
-    public ReservaDTO buscarReservaPorId(Long id) {
+    public ReservaDTO buscarReservaPorId(Integer id) {
         ReservaEntity reservaEntity = reservaRepository.findById(id)
                 .orElseThrow(() -> new ReservaNaoEncontradaException("A reserva com o ID " + id + " não existe."));
 
@@ -90,8 +93,9 @@ public class ReservaService {
                 novaReserva.getDataCheckOut()
         );
 
-        if (reservaEntity == null)
+        if (reservaEntity == null){
             return false;
+        }
 
         return reservaEntity.getStatusReserva() != StatusReservaEnum.EM_ANDAMENTO;
     }
